# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Propolis Fulfillment API",
        "description": "Dokumentasi resmi layanan backend Propolis. Semua response menggunakan format {success, data, message}.",
        "contact": {
            "name": "Tim Propolis",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Root"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Daftar produk",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status produk (aktif/nonaktif)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "kategori_id",
                        "in": "query",
                        "description": "Filter berdasarkan kategori",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil memuat produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tambah produk baru",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kategori_id",
                                    "sku",
                                    "nama_produk",
                                    "harga_ecer",
                                    "status"
                                ],
                                "properties": {
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "sku": {
                                        "type": "string"
                                    },
                                    "nama_produk": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "harga_ecer": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "gambar": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "nonaktif"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Perbarui produk",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "sku": {
                                        "type": "string"
                                    },
                                    "nama_produk": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "harga_ecer": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    },
                                    "gambar": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "nonaktif"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk diperbarui"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{productId}/price-tiers": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tambahkan harga tingkat",
                "operationId": "7d267b2b371593603b6b87770d35e6c2",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "min_jumlah",
                                    "harga_total"
                                ],
                                "properties": {
                                    "min_jumlah": {
                                        "type": "integer"
                                    },
                                    "max_jumlah": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "harga_total": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "label": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Harga tingkat ditambahkan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/batch-stock": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Laporan stok per batch",
                "operationId": "de0ae838aa8832b64efb6edf8e06187d",
                "responses": {
                    "200": {
                        "description": "Berhasil memuat laporan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/batch-sales": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Laporan penjualan per batch",
                "operationId": "622d41b36b8c0d52a26b086cf34de87c",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Tanggal awal (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Tanggal akhir (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil memuat laporan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Masukkan token Sanctum dengan format \"Bearer {token}\"."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}